refactor: restructure orders-list and order-details components to modular architecture
BREAKING CHANGE: Refactor remaining order components from inline templates to modular structure

Changes:
- Refactor orders-list component into separate files (HTML, CSS, TS, index)
- Refactor order-details component into separate files (HTML, CSS, TS, index)
- Update route imports to use new modular component paths
- Remove old monolithic component files
- Maintain all existing functionality and logic

Files:
- apps/frontend/src/app/private/modules/store/orders/list/orders-list/
- apps/frontend/src/app/private/modules/store/orders/details/order-details/
- apps/frontend/src/app/routes/private/store_admin.routes.ts

Components now follow consistent modular architecture:
- order-stats (✅ previously refactored)
- orders-list (✅ newly refactored)
- order-details (✅ newly refactored)

All components maintain their original functionality while improving:
- Code maintainability and separation of concerns
- Developer experience with dedicated HTML/CSS files
- Consistency across the entire orders module
