--- apps/frontend/src/app/private/modules/store/orders/details/order-details.component.ts
@@ -1,50 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ActivatedRoute } from '@angular/router';
-
-@Component({
-  selector: 'app-order-details',
-  standalone: true,
-  imports: [CommonModule],
-  template: `
-    <div class="p-6">
-      <div class="mb-6">
-        <h1 class="text-3xl font-bold text-gray-900 mb-2">Order Details</h1>
-        <p class="text-gray-600">View detailed information about this order</p>
-      </div>
-
-      <div class="bg-white rounded-lg shadow-sm border p-8">
-        <div class="text-center">
-          <div class="inline-flex items-center justify-center w-16 h-16 bg-blue-100 rounded-full mb-4">
-            <svg class="w-8 h-8 text-blue-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"></path>
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M2.458 12C3.732 7.943 7.523 5 12 5c4.478 0 8.268 2.943 9.542 7-1.274 4.057-5.064 7-9.542 7-4.477 0-8.268-2.943-9.542-7z"></path>
-            </svg>
-          </div>
-          <h2 class="text-xl font-semibold text-gray-900 mb-2">Order Details</h2>
-          <p class="text-gray-600 max-w-md mx-auto">
-            Order ID: #{{orderId || 'Loading...' }}
-          </p>
-          <p class="text-gray-600 max-w-md mx-auto mt-2">
-            Order details view is under development. You will be able to view complete order information here.
-          </p>
-        </div>
-      </div>
-    </div>
-  `,
-  styles: [`
-    :host {
-      display: block;
-      width: 100%;
-    }
-  `]
-})
-export class OrderDetailsComponent {
-  orderId: string | null = null;
-
-  constructor(private route: ActivatedRoute) {
-    this.route.paramMap.subscribe(params => {
-      this.orderId = params.get('id');
-    });
-  }
-}
