+++ apps/frontend/src/app/private/modules/store/orders/details/order-details/index.ts
@@ -1,0 +1,1 @@
+export { OrderDetailsComponent } from './order-details.component';
