+++ apps/frontend/src/app/private/modules/store/orders/details/order-details/order-details.component.css
@@ -1,0 +1,4 @@
+:host {
+  display: block;
+  width: 100%;
+}
