+++ apps/frontend/src/app/private/modules/store/orders/list/orders-list/index.ts
@@ -1,0 +1,1 @@
+export { OrdersListComponent } from './orders-list.component';
