+++ apps/frontend/src/app/private/modules/store/orders/list/orders-list/orders-list.component.css
@@ -1,0 +1,4 @@
+:host {
+  display: block;
+  width: 100%;
+}
