+++ apps/frontend/src/app/private/modules/store/orders/list/orders-list/orders-list.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-orders-list',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './orders-list.component.html',
+  styleUrls: ['./orders-list.component.css'],
+})
+export class OrdersListComponent {}
